refactor: introduce http client
