+++ granulate_utils/config_feeder/client/client.py
@@ -2,26 +2,20 @@ import asyncio
 import json
 import logging
 from collections import defaultdict
-from typing import Any, DefaultDict, Dict, Optional, Union, cast
-
-from pydantic import BaseModel
-from requests import Session
-from requests.exceptions import ConnectionError, JSONDecodeError
+from typing import DefaultDict, Dict, List, Optional, Union
 
 from granulate_utils.config_feeder.client.bigdata import get_node_info
-from granulate_utils.config_feeder.client.exceptions import APIError, ClientError
+from granulate_utils.config_feeder.client.collector import ConfigFeederCollector
+from granulate_utils.config_feeder.client.exceptions import ClientError
+from granulate_utils.config_feeder.client.http_client import HttpClient
 from granulate_utils.config_feeder.client.models import CollectionResult, ConfigType
 from granulate_utils.config_feeder.client.yarn.collector import YarnConfigCollector
 from granulate_utils.config_feeder.client.yarn.models import YarnConfig
-from granulate_utils.config_feeder.core.errors import raise_for_code
 from granulate_utils.config_feeder.core.models.aggregation import CreateNodeConfigsRequest, CreateNodeConfigsResponse
 from granulate_utils.config_feeder.core.models.cluster import ClusterCreate, CreateClusterRequest, CreateClusterResponse
 from granulate_utils.config_feeder.core.models.collection import CollectorType
 from granulate_utils.config_feeder.core.models.node import CreateNodeRequest, CreateNodeResponse, NodeCreate, NodeInfo
 from granulate_utils.config_feeder.core.models.yarn import NodeYarnConfigCreate
-
-DEFAULT_API_SERVER_ADDRESS = "https://api.granulate.io/config-feeder/api/v1"
-DEFAULT_REQUEST_TIMEOUT = 3
 
 
 class ConfigFeederClient:
@@ -34,19 +28,19 @@ class ConfigFeederClient:
         server_address: Optional[str] = None,
         yarn: bool = True,
         collector_type=CollectorType.SAGENT,
+        collectors: List[ConfigFeederCollector] = [],
     ) -> None:
         if not token or not service:
             raise ClientError("Token and service must be provided")
         self.logger = logger
-        self._token = token
         self._service = service
         self._cluster_id: Optional[str] = None
         self._collector_type = collector_type
-        self._server_address: str = server_address.rstrip("/") if server_address else DEFAULT_API_SERVER_ADDRESS
         self._is_yarn_enabled = yarn
         self._yarn_collector = YarnConfigCollector(logger=logger)
         self._last_hash: DefaultDict[ConfigType, Dict[str, str]] = defaultdict(dict)
-        self._init_api_session()
+        self._collectors = collectors
+        self._http_client = HttpClient(token, server_address)
 
     def collect(self) -> None:
         if (node_info := get_node_info(self.logger)) is None:
@@ -90,7 +84,7 @@ class ConfigFeederClient:
 
         node_id = self._register_node(collection_result.node)
         self.logger.info(f"sending configs for node {external_id}")
-        response = CreateNodeConfigsResponse(**self._api_request("POST", f"/nodes/{node_id}/configs", request))
+        response = CreateNodeConfigsResponse(**self._http_client.request("POST", f"/nodes/{node_id}/configs", request))
 
         if response.yarn_config is not None:
             assert request.yarn_config is not None
@@ -110,7 +104,9 @@ class ConfigFeederClient:
             ),
             allow_existing=True,
         )
-        response = CreateNodeResponse(**self._api_request("POST", f"/clusters/{self._cluster_id}/nodes", request))
+        response = CreateNodeResponse(
+            **self._http_client.request("POST", f"/clusters/{self._cluster_id}/nodes", request)
+        )
         return response.node.id
 
     def _register_cluster(self, node_info: NodeInfo) -> None:
@@ -126,7 +122,7 @@ class ConfigFeederClient:
             ),
             allow_existing=True,
         )
-        response = CreateClusterResponse.parse_obj(self._api_request("POST", "/clusters", request))
+        response = CreateClusterResponse.parse_obj(self._http_client.request("POST", "/clusters", request))
         self._cluster_id = response.cluster.id
 
     def _get_configs_request(self, configs: CollectionResult) -> Optional[CreateNodeConfigsRequest]:
@@ -148,35 +144,3 @@ class ConfigFeederClient:
         return NodeYarnConfigCreate(
             collector_type=self._collector_type, config_json=json.dumps(configs.yarn_config.config)
         )
-
-    def _api_request(
-        self,
-        method: str,
-        path: str,
-        request_data: Optional[BaseModel] = None,
-        timeout: float = DEFAULT_REQUEST_TIMEOUT,
-    ) -> Dict[str, Any]:
-        try:
-            resp = self._session.request(
-                method,
-                f"{self._server_address}{path}",
-                json=request_data.dict() if request_data else None,
-                timeout=timeout,
-            )
-            if resp.ok:
-                return cast(Dict[str, Any], resp.json())
-            try:
-                res = resp.json()
-                if "detail" in res:
-                    raise APIError(res["detail"], path, resp.status_code)
-                error = res["error"]
-                raise_for_code(error["code"], error["message"])
-                return cast(Dict[str, Any], res)
-            except (KeyError, JSONDecodeError):
-                raise APIError(resp.text or resp.reason, path, resp.status_code)
-        except ConnectionError:
-            raise ClientError(f"could not connect to {self._server_address}")
-
-    def _init_api_session(self) -> None:
-        self._session = Session()
-        self._session.headers.update({"Accept": "application/json", "GProfiler-API-Key": self._token})
