+++ granulate_utils/config_feeder/client/collector.py
@@ -1,0 +1,2 @@
+class ConfigFeederCollector:
+    pass
