+++ tests/granulate_utils/config_feeder/fixtures/api.py
@@ -7,7 +7,7 @@ from unittest.mock import patch
 from requests_mock.mocker import Mocker
 from requests_mock.request import _RequestObjectProxy
 
-from granulate_utils.config_feeder.client.client import DEFAULT_API_SERVER_ADDRESS
+from granulate_utils.config_feeder.client.http_client import DEFAULT_API_SERVER_ADDRESS
 from granulate_utils.config_feeder.core.models.cluster import BigDataPlatform, CloudProvider
 from granulate_utils.config_feeder.core.models.node import NodeInfo
